+++ src/app/reducer/index.js
@@ -1,0 +1,16 @@
+import { combineReducers } from 'redux';
+import moviesReducer from './movies.js';
+import movieReducer from './movie.js';
+import actorsReducer from './actors.js';
+import actorReducer from './actor.js';
+import scoreReducer from './score.js';
+
+const rootReducer = combineReducers({
+	movies: moviesReducer,
+	movie: movieReducer,
+	actors: actorsReducer,
+	actor: actorReducer,
+	score: scoreReducer
+});
+
+export default rootReducer;