+++ src/app/reducer/score.js
@@ -1,0 +1,21 @@
+const initialState = []
+
+const scoreReducer = (state=initialState,action) =>{
+    switch(action.type){
+        case "UPDATE_SCORE": {
+            return{
+                ...state,
+                score: action.payload,
+                status:'update'
+            }
+        }
+        case "CLEAR_SCORE": {
+            return state = [];
+        }
+        default: {
+            return state
+        }
+    }
+}
+
+export default scoreReducer;